pyproject.toml
[project]
name = "algobet"
version = "0.1.0"
description = "Football match database and OddsPortal scraper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "optuna>=3.4.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tabulate>=0.9.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
    "apscheduler>=3.10.0",
    "croniter>=2.0.0",
    "dependency-injector>=4.41.0",
    "nest-asyncio>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "mypy>=1.7.0",
]

[project.scripts]
algobet = "algobet.cli.dev_tools:cli"
algobet-dev = "algobet.cli.dev_tools:cli"
algobet-scheduler = "algobet.scheduler.worker:cli"
algobet-runner = "algobet.cli.scheduled_runner:cli"

[tool.ruff]
target-version = "py310"
line-length = 88
indent-width = 4
exclude = [
    ".git",
    ".git-rewrite",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pypackages__",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # Pyflakes
    "I",    # isort
    "W",    # pycodestyle warnings
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
]
ignore = [
    "B008",  # FastAPI uses function calls in argument defaults (Query, Depends)
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/conftest.py" = ["E402"]  # Imports not at top for monkeypatching
"algobet/scraper.py" = ["E501", "W291", "W293"]  # JavaScript strings inside Python

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["algobet"]
force-single-line = false
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "types-croniter>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
ignore_missing_imports = false
mypy_path = "stubs"

# Ignore missing imports for libraries without stubs
[[tool.mypy.overrides]]
module = [
    "playwright.*",
    "psycopg2.*",
    "psycopg2_binary.*",
    "tabulate.*",
    "croniter.*",
]
ignore_missing_imports = true

# SQLAlchemy plugin settings for 2.0 style
[[tool.mypy.overrides]]
module = ["algobet.models", "algobet.database"]
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["sqlalchemy.*"]
ignore_missing_imports = false

# Pydantic plugin settings for v2
[[tool.mypy.overrides]]
module = ["pydantic.*"]
ignore_missing_imports = false

# Tests - disable strict checking
[[tool.mypy.overrides]]
module = ["tests.*"]
strict = false
